+++ config/__init__.py
@@ -1,1 +1,2 @@
 from .templates import TEMPLATES_DIR as templates
+from .paths import TRELLO_TOKEN, TRELLO_KEY, CSV_DIR